fix: parser-parity 測試拆成零瀏覽器不變量檢查＋e2e 雙 parser 比對
原本的 tests/data/parser-parity.test.ts 在 vitest 裡直接 chromium.launch()，本機因為之前
跑過 E2E、~/.cache/ms-playwright/ 剛好留有瀏覽器而測試能過，但 npm test 跑在 ci.yml 的
verify job，該 job 從頭到尾沒有 npx playwright install（只有 e2e job 裝了瀏覽器），兩個
job 又跑在各自獨立的 runner 上——乾淨的 verify job 會直接炸「Executable doesn't exist」，
擋掉每一個 PR。

改成兩支各司其職的測試：
- tests/data/parser-parity.test.ts（vitest）改成純字串掃描 data/dice-tree.svg，零瀏覽器
  相依，驗「換行不變量本身有沒有被破壞」（data-* 屬性值內字面換行數 = 0、<title> 內含字面
  換行的節點數 = 69，皆為 Task 1 已核實的實際值）。
- 新增 tests/e2e/parser-parity.spec.ts（Playwright，只在裝了瀏覽器的 e2e job 跑），
  用 Chromium DOMParser 與 linkedom 雙 parser 比對，證明「為什麼」這個不變量重要。

CLAUDE.md 同步更新測試環境段落，說明這個切分與 verify job 沒裝瀏覽器的事實，避免日後
把需要瀏覽器的測試誤放進 npm test 的範圍。
